+++ hosts/echo/home.nix
@@ -1,1 +1,5 @@
-{ ... }: { }
+{ ... }:
+
+{
+  programs.firefox.enable = false;
+}