+++ modules/base.nix
@@ -257,9 +257,6 @@ in {
         prezto = {
           enable = true;
 
-          # https://github.com/nix-community/home-manager/issues/2255
-          caseSensitive = true;
-
           pmoduleDirs = [ "${pkgs.zsh-you-should-use}/share/zsh/plugins" ];
 
           pmodules = [